src/requirements.txt
streamlit
openai-whisper
torch
librosa
soundfile
noisereduce
pyloudnorm
scikit-learn
nltk
textblob
textstat
plotly
wordcloud
matplotlib
requests